Fix Carbon hotkey byte-count type

+++ Sources/HotKeyManager.swift
@@ -46,7 +46,7 @@ final class HotKeyManager {
                     EventParamName(kEventParamDirectObject),
                     EventParamType(typeEventHotKeyID),
                     nil,
-                    UInt32(MemoryLayout<EventHotKeyID>.size),
+                    MemoryLayout<EventHotKeyID>.size,
                     nil,
                     &hotKeyID
                 )
